Moved stuff around and made list tests pass
