+++ packages/babel-plugin/src/tests/transform.ts
@@ -18,7 +18,7 @@
 
 import * as babel from "@babel/core";
 
-import plugin from "../index";
+import plugin from "../plugin";
 
 export type TransformOptions = {
   source: string;
